refactor: 타스 리팩토링
