+++ memberPage/signInUp/package.json
@@ -14,7 +14,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.18.0",
-    "styled-components": "^6.1.1"
+    "styled-components": "^6.1.4"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
