+++ memberPage/signInUp/src/App.tsx
@@ -1,6 +1,7 @@
 import { RouterProvider } from "react-router-dom";
 import GlobalStyles from "./styles/Globalstyle";
 import Router from "./components/Router";
+import React from "react";
 
 function App() {
   return (
