+++ memberPage/signInUp/src/page/SignUp.tsx
@@ -12,6 +12,7 @@ import {
 import { useNavigate } from "react-router-dom";
 
 import axios from "axios";
+import React from "react";
 
 const SignUp = () => {
   const [activeSignUp, setActiveSignUp] = useState(false);
